feat(proof): rework base-proof-host
