+++ crates/proof/host/Cargo.toml
@@ -33,21 +33,20 @@ revm = { workspace = true, optional = true, features = ["std", "c-kzg", "secp256
 # Optional - disk
 rocksdb = { workspace = true, optional = true, features = ["snappy", "bindgen-runtime"] }
 
-# Optional - single
-serde = { workspace = true, optional = true }
-ark-ff = { workspace = true, optional = true }
-alloy-eips = { workspace = true, optional = true }
-alloy-network = { workspace = true, optional = true }
-base-protocol = { workspace = true, optional = true }
-base-alloy-network = { workspace = true, optional = true }
-base-proof-std-fpvm = { workspace = true, optional = true }
-base-consensus-providers = { workspace = true, optional = true }
-alloy-rpc-types-beacon = { workspace = true, optional = true }
-base-proof = { workspace = true, optional = true, features = ["std"] }
-alloy-rpc-types = { workspace = true, optional = true, features = ["debug"] }
-clap = { workspace = true, optional = true, features = ["derive", "env"] }
-base-consensus-genesis = { workspace = true, optional = true, features = ["serde", "std"] }
-base-alloy-rpc-types-engine = { workspace = true, optional = true, features = ["serde", "std"] }
+serde.workspace = true
+ark-ff.workspace = true
+alloy-eips.workspace = true
+alloy-network.workspace = true
+base-protocol.workspace = true
+base-alloy-network.workspace = true
+base-proof-std-fpvm.workspace = true
+base-consensus-providers.workspace = true
+alloy-rpc-types-beacon.workspace = true
+base-proof = { workspace = true, features = ["std"] }
+alloy-rpc-types = { workspace = true, features = ["debug"] }
+clap = { workspace = true, features = ["derive", "env"] }
+base-consensus-genesis = { workspace = true, features = ["serde", "std"] }
+base-alloy-rpc-types-engine = { workspace = true, features = ["serde", "std"] }
 
 [dev-dependencies]
 proptest.workspace = true
@@ -56,24 +55,5 @@ proptest.workspace = true
 workspace = true
 
 [features]
-default = []
 disk = [ "dep:rocksdb" ]
 precompiles = [ "dep:revm" ]
-single = [
-	"dep:alloy-eips",
-	"dep:alloy-network",
-	"dep:alloy-rpc-types",
-	"dep:alloy-rpc-types-beacon",
-	"dep:ark-ff",
-	"dep:base-alloy-network",
-	"dep:base-alloy-rpc-types-engine",
-	"dep:base-consensus-genesis",
-	"dep:base-consensus-providers",
-	"dep:base-proof",
-	"dep:base-proof-std-fpvm",
-	"dep:base-protocol",
-	"dep:clap",
-	"dep:serde",
-	"disk",
-	"precompiles",
-]
