+++ crates/proof/host/src/backend/mod.rs
@@ -1,12 +1,8 @@
-//! Backend for the preimage server.
-
 mod offline;
 pub use offline::OfflineHostBackend;
 
-#[cfg(feature = "single")]
 mod online;
-#[cfg(feature = "single")]
-pub use online::{HintHandler, OnlineHostBackend, OnlineHostBackendCfg};
+pub use online::OnlineHostBackend;
 
 mod util;
 pub use util::store_ordered_trie;
