--- crates/proof/host/src/single/mod.rs
@@ -1,10 +1,0 @@
-//! This module contains the single-chain mode for the host.
-
-mod cfg;
-pub use cfg::{SingleChainHost, SingleChainHostError, SingleChainProviders};
-
-mod local_kv;
-pub use local_kv::SingleChainLocalInputs;
-
-mod handler;
-pub use handler::{SingleChainHintHandler, parse_blob_hint};
